репликация метрики 0,370/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126

torch==2.9.1
torchvision==0.24.1
torchaudio==2.9.1
absl-py==2.4.0
accelerate==1.12.0
adagio==0.2.6
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.3
annotated-types==0.7.0
asttokens==2.4.1
attrs==25.4.0
autogluon.common==1.5.0
autogluon.core==1.5.0
autogluon.features==1.5.0
autogluon.tabular==1.5.0
autogluon.timeseries==1.5.0
boto3==1.42.44
botocore==1.42.44
catboost==1.2.8
certifi==2026.1.4
charset-normalizer==3.4.4
chronos-forecasting==2.2.2
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
comm==0.2.0
contourpy==1.3.3
coreforecast==0.0.16
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
einops==0.8.2
executing==2.0.1
autogluon.tabular[fastai]==1.5.0
filelock==3.20.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
fugue==0.9.6
gluonts==0.16.2
google_search_results==2.4.2
graphviz==0.21
greenlet==3.3.1
grpcio==1.78.0
holidays==0.90
huggingface_hub==0.36.2
idna==3.11
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.9
lightgbm==4.6.0
lightning==2.5.6
lightning-utilities==0.15.2
llvmlite==0.46.0
Mako==1.3.10
Markdown==3.10.1
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.1.6
mlforecast==0.14.0
mpmath==1.3.0
multidict==6.7.1
narwhals==2.16.0
nest-asyncio==1.5.8
networkx==3.6.1
numba==0.63.1
numpy==2.1.3
optuna==4.7.0
orjson==3.11.7
pandas==2.3.3
parso==0.8.3
patsy==1.0.2
peft==0.17.1
pillow==12.1.0
platformdirs==4.1.0
plotly==6.5.2
prompt-toolkit==3.0.41
propcache==0.4.1
protobuf==6.33.5
psutil==5.9.6
pure-eval==0.2.2
pyarrow==20.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.17.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==25.1.2
ray==2.50.0
regex==2026.1.15
requests==2.32.5
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentencepiece==0.2.1
six==1.17.0
SQLAlchemy==2.0.46
stack-data==0.6.3
statsforecast==2.0.1
statsmodels==0.14.6
tsfresh>=0.20.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.2
toolz==0.12.1
torchmetrics==1.8.2
tornado==6.4
tqdm==4.67.3
traitlets==5.14.0
transformers==4.57.6
triad==1.0.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
utilsforecast==0.2.11
wcwidth==0.2.12
Werkzeug==3.1.5
window_ops==0.0.15
xgboost==3.1.3
yarl==1.22.0